+++ R/layout.R
@@ -3434,8 +3434,7 @@ layout_components <- function(graph, layout = NULL, ...) {
 #' @export
 #' @keywords internal
 layout.spring <- function(graph, ...) {
-  lifecycle::deprecate_warn("2.1.0", "layout.spring()", "layout_with_fr()")
-  layout_with_fr(graph)
+  lifecycle::deprecate_stop("2.1.0", "layout.spring()", "layout_with_fr()")
 }
 
 #' SVD layout, this was removed from igraph
@@ -3452,8 +3451,7 @@ layout.spring <- function(graph, ...) {
 #' @keywords internal
 #' @export
 layout.svd <- function(graph, ...) {
-  lifecycle::deprecate_warn("2.1.0", "layout.svd()", "layout_with_fr()")
-  layout_with_fr(graph)
+  lifecycle::deprecate_stop("2.1.0", "layout.svd()", "layout_with_fr()")
 }
 
 #' Grid Fruchterman-Reingold layout, this was removed from igraph
@@ -3471,12 +3469,11 @@ layout.svd <- function(graph, ...) {
 #' @keywords internal
 #' @export
 layout.fruchterman.reingold.grid <- function(graph, ...) {
-  lifecycle::deprecate_warn(
+  lifecycle::deprecate_stop(
     "2.1.0",
     "layout.fruchterman.reingold.grid()",
     "layout_with_fr()"
   )
-  layout_with_fr(graph)
 }
 
 #' The DrL graph layout generator